+++ src/xiaomi_miwifi/__init__.py
@@ -16,4 +16,4 @@ __all__ = [
     "SUPPORTED_ROUTERS",
     "friendly_model",
 ]
-__version__ = "0.1.0"
+__version__ = "0.2.0"